Parser starting to look like a parser. also a journal entry.
Expression::Expressions renamed to Expression::Application to express better its purpose in the syntax.

Removed the Sink (and the generics for it) from the parser as the error handling changed to immediately crashing on error. Since the parser only outputs the final expression, differents threads and a pipeline between them would be unnecessary.

Almost all the parts of the parser are done except closing bracket handling (and the handling of when the tokens run out).

Also rambled some stuff about what I've been doing in the journal.